requirements.txt
streamlit
langchain_core
langchain_community
langchain_ollama
langchain-unstructured
pypdf
faiss-cpu
chroma
unstructured
unstructured[all-docs]
dotenv


# Web 應用框架
streamlit>=1.31.0

# LangChain 核心套件
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0  
langchain-unstructured>=0.1.0
langchain-text-splitters>=0.2.0

# 向量儲存與檢索
faiss-cpu>=1.8.0

# 環境變數管理
python-dotenv>=1.0.0

# 文件處理與日誌
unstructured>=0.12.0
pillow>=10.0.0

# 其他依賴
requests>=2.31.0
numpy>=1.26.0

psycopg2-binary

langsmith